ExternalPaginationTable custom class
