+++ jdi-light-vuetify-tests/src/main/java/io/github/com/pages/DataTablesPage.java
@@ -4,6 +4,7 @@ import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.html.elements.common.Button;
 import com.epam.jdi.light.vuetify.elements.complex.tables.DataTable;
 import io.github.com.custom.cards.NewItemCard;
+import io.github.com.custom.tables.ExternalPaginationTable;
 import io.github.com.custom.tables.ExternalSortingTable;
 
 public class DataTablesPage extends VuetifyPage {
@@ -66,7 +67,7 @@ public class DataTablesPage extends VuetifyPage {
     public static DataTable expandableRowsTable;
 
     @UI("#ExternalPaginationTable")
-    public static DataTable externalPaginationTable;
+    public static ExternalPaginationTable externalPaginationTable;
 
     @UI("#ExternalSortingTable")
     public static ExternalSortingTable externalSortingTable;
